Factor out root routes from app.js to own file
